Uploading files

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -7,15 +7,15 @@
  */
 void print_python_list_info(PyObject *p)
 {
-	int i;
+	long int i;
 	long int size = PyList_Size(p);
-	long int allocated = ((PyListObject*)(p))->allocated;
-	
+	long int allocated = ((PyListObject *)(p))->allocated;
+
 	printf("[*] Size of the Python List = %li\n", size);
 	printf("[*] Allocated = %li\n", allocated);
 	for (i = 0; i < size; i++)
-		{
-			temp = PyList_GetItem(p, i);
-			printf("Element %d: %s\n", i, Py_TYPE(temp)->tp_name);
-		}
+	{
+		temp = PyList_GetItem(p, i);
+		printf("Element %d: %s\n", i, Py_TYPE(temp)->tp_name);
+	}
 }
